MUST_REBASE [VPG]: adf/intel: Memory management for VLV
VLV uses GEM memory manager. Complete implementation of memory will be
done when we enable FB driver. For now Gralloc allocates memory using
DRM prime interface which gives DMA Buf handle over GEM buffers.

So just
- do proper buffer mapping for use
- Disable out default ADF memory initialization and validations if VLV is
  enabled at compile time
- Do not include intel_adf_mm.c for compilation

MUST_REBASE: Since Google ADF framework is not in upstream and also
we cannot have two display drivers, other one being i915, we need to
work on i915 and ADF convergence path before it can be upstreamed

For: GMINL-3547

+++ drivers/video/adf/intel/include/core/intel_dc_config.h
@@ -105,7 +105,7 @@ struct intel_buffer {
 	u32 w;
 	u32 h;
 	u32 format;
-	u32 gtt_offset_in_pages;
+	unsigned long gtt_offset_in_pages;
 	u32 stride;
 };
